fixes the updating

+++ core/src/main/scala/nl/jappieklooster/gdx/mapstare/akka/WorldUpdateActor.scala
@@ -61,7 +61,7 @@ class WorldUpdateActor extends Actor with Logging{
 				// the slow parts.
 				if(difference > WorldUpdateActor.minimumReasonableUpdate) {
 					lastTime += difference
-					val tick = GameTick(difference/1000)
+					val tick = GameTick((difference.toDouble/1000).toFloat)
 
 					updater.update(tick)
